Updated example to use the functional API instead of sequential
I updated the code in the example so that now I can use the functional API to build the keras model instead of the sequential function. To achieve this, I introduced a new layer in the model that converts the input from the feature columns into something that can be fed into the first layer of the neural network without warnings of data incompatibility.